+++ APISistemaTickets/Data/Services/ICommentService.cs
@@ -5,8 +5,8 @@ namespace APISistemaTickets.Data.Services;
 public interface ICommentService
 {
     Task<IEnumerable<Comment>?> GetAll();
-    Task<Comment?> GetById(int id);
+    Task<Comment?> GetById(long id);
     Task<Comment> Create(Comment comment);
     Task<Comment> Update(Comment comment);
-    Task Delete(int id);
+    Task Delete(long id);
 }
